requirements.txt
aiohttp==3.8.6
aiosignal==1.2.0
astroid==2.11.7
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
datasets==2.4.0
dill==0.3.4
filelock==3.4.1
Flask==2.0.3
frozenlist==1.2.0
fsspec==2022.1.0
huggingface-hub==0.4.0
idna==3.6
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
lxml==5.2.0
MarkupSafe==2.0.1
mccabe==0.7.0
multidict==5.2.0
multiprocess==0.70.12.2
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.0
pyarrow==6.0.1
pylint==2.13.9
pyparsing==3.0.7
pyspark==2.4.3
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
responses==0.17.0
six==1.16.0
tomli==1.2.3
torch==1.10.2
tqdm==4.64.1
typed-ast==1.5.5
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
wrapt==1.16.0
xxhash==3.2.0
yarl==1.7.2
zipp==3.6.0
